# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HedgeAction": {
        "description": "Portfolio-level action a hypothesis or reassessment proposes.\n\nCovers both a fresh hedge and the adaptive-lifecycle outcomes in BRD §28\n(``MAINTAIN`` / ``INCREASE`` / ``DECREASE`` / ``REMOVE`` / ``REPLACE`` /\n``NO_TRADE``).",
        "enum": [
          "NEW_HEDGE",
          "INCREASE",
          "DECREASE",
          "MAINTAIN",
          "REMOVE",
          "REPLACE",
          "NO_TRADE"
        ],
        "title": "HedgeAction",
        "type": "string"
      },
...
      "HedgeMetrics": {
        "additionalProperties": false,
        "description": "Quantitative summary of a hypothesis — every number sourced from ``quant/``.",
        "properties": {
          "breakevens": {
            "items": {
              "type": "number"
            },
            "title": "Breakevens",
            "type": "array"
          },
          "cost_pct_of_portfolio": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Pct Of Portfolio"
          },
          "downside_protection_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downside Protection Pct"
          },
          "hedge_ratio": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hedge Ratio"
          },
          "max_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Loss"
          },
          "net_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Delta"
          },
          "net_gamma": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Gamma"
          },
          "net_theta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Theta"
          },
          "net_vega": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Vega"
          }
        },
        "title": "HedgeMetrics",
        "type": "object"
      },
...
      "OptionLeg": {
        "additionalProperties": false,
        "description": "One option leg of a hedge structure.",
        "properties": {
          "expiration": {
            "format": "date",
            "title": "Expiration",
            "type": "string"
          },
          "limit_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "occ_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "full OCC option identifier when resolved",
            "title": "Occ Symbol"
          },
          "quantity": {
            "description": "number of contracts (always positive)",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "right": {
            "$ref": "#/components/schemas/OptionRight"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "strike": {
            "exclusiveMinimum": 0.0,
            "title": "Strike",
            "type": "number"
          },
          "underlying": {
            "title": "Underlying",
            "type": "string"
          }
        },
        "required": [
          "underlying",
          "right",
          "side",
          "strike",
          "expiration",
          "quantity"
        ],
        "title": "OptionLeg",
        "type": "object"
      },
      "OptionRight": {
        "description": "Option contract right.",
        "enum": [
          "CALL",
          "PUT"
        ],
        "title": "OptionRight",
        "type": "string"
      },
...
      "OrderSide": {
        "description": "Direction of a leg or position.",
        "enum": [
          "BUY",
          "SELL"
        ],
        "title": "OrderSide",
        "type": "string"
      },
      "PayoffPoint": {
        "additionalProperties": false,
        "description": "One ``(underlying price, portfolio P&L)`` sample of the payoff curve.",
        "properties": {
          "pnl": {
            "title": "Pnl",
            "type": "number"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "number"
          }
        },
        "required": [
          "price",
          "pnl"
        ],
        "title": "PayoffPoint",
        "type": "object"
      },
...
      "StrategyDecisionOut": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "alternatives": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternatives"
          },
          "comparison": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "selected_hypothesis_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Hypothesis Id"
          }
        },
        "required": [
          "id",
          "cycle_id",
          "action",
          "rationale"
        ],
        "title": "StrategyDecisionOut",
        "type": "object"
      },
      "StrategyHypothesis": {
        "additionalProperties": false,
        "description": "A single strategy proposal from one Strategy Agent (BRD §16).",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/HedgeAction"
          },
          "cost": {
            "description": "absolute premium/debit in account currency",
            "minimum": 0.0,
            "title": "Cost",
            "type": "number"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "hedge_metrics": {
            "$ref": "#/components/schemas/HedgeMetrics"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/OptionLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "liquidity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liquidity"
          },
          "payoff_profile": {
            "items": {
              "$ref": "#/components/schemas/PayoffPoint"
            },
            "title": "Payoff Profile",
            "type": "array"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "rejection_conditions": {
            "description": "conditions under which this proposal stops being valid",
            "items": {
              "type": "string"
            },
            "title": "Rejection Conditions",
            "type": "array"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "why the agent rejected its own family (viable is False)",
            "title": "Rejection Reason"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "title": "Risks",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategyType"
          },
          "tradeoffs": {
            "items": {
              "type": "string"
            },
            "title": "Tradeoffs",
            "type": "array"
          },
          "viable": {
            "title": "Viable",
            "type": "boolean"
          }
        },
        "required": [
          "cycle_id",
          "strategy",
          "action",
          "viable",
          "cost",
          "rationale"
        ],
        "title": "StrategyHypothesis",
        "type": "object"
      },
      "StrategyHypothesisOut": {
        "properties": {
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "legs": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legs"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "strategy_type": {
            "title": "Strategy Type",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cycle_id",
          "strategy_type",
          "verdict"
        ],
        "title": "StrategyHypothesisOut",
        "type": "object"
      },
      "StrategyType": {
        "description": "The four hedge strategy families in the MVP (BRD §15).",
        "enum": [
          "PROTECTIVE_PUT",
          "PUT_SPREAD",
          "COLLAR",
          "NO_HEDGE"
        ],
        "title": "StrategyType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Autonomous Adaptive Portfolio Hedge Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/strategy/decision": {
      "get": {
        "description": "The decision for ``cycle_id`` (or the most recent one when omitted).",
        "operationId": "strategy_decision_strategy_decision_get",
        "parameters": [
          {
            "description": "the strategy-evaluation cycle to read",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "the strategy-evaluation cycle to read",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strategy Decision",
        "tags": [
          "strategy-readback"
        ]
      }
    },
...
    "/strategy/hypotheses": {
      "get": {
        "description": "All hypotheses in insertion order, optionally filtered to one cycle.\n\nRejected / NOT_VIABLE hypotheses are included — the rejection reasoning is\npart of the decision trail.",
        "operationId": "strategy_hypotheses_strategy_hypotheses_get",
        "parameters": [
          {
            "description": "restrict to one strategy-evaluation cycle",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "restrict to one strategy-evaluation cycle",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyHypothesisOut"
                  },
                  "title": "Response Strategy Hypotheses Strategy Hypotheses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strategy Hypotheses",
        "tags": [
          "strategy-readback"
        ]
      }
    },
    "/strategy/hypotheses/{hypothesis_id}": {
      "get": {
        "description": "A single StrategyHypothesis proposal (P4-DB-3 / #110).",
        "operationId": "get_strategy_hypothesis_strategy_hypotheses__hypothesis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "hypothesis_id",
            "required": true,
            "schema": {
              "title": "Hypothesis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyHypothesis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy Hypothesis",
        "tags": [
          "strategy-readback"
        ]
      }
    },
...
  }
}